Initial commit of Kuehlmak keyboard layout generator
This will be a Rust reimplementation of an earlier Python project. This
is also my entry into Rust programming.

This initial commit includes the first version of the text_stats module
that analyzes the frequency of symbols, bigrams and 3-grams in input
texts, including some unit tests. Deserialization from JSON is still
missing.

The basic layout is a binary and a library. The binary is just a dummy
for now. Later there will be multiple binaries sharing common
functionality from the library to implement different tools, e.g.:

 * A text_stats analyzer for large bodies of text that saves JSON
 * A standalone keyboard layout evaluation
 * A layout optimizer using simulated annealing

There is also a benches module using the bencher crate to benchmark
different components later. For now there are some benchmarks for the
n-gram lookup methods in TextStats.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>