eval: Implement printing of scores
Added a trait EvalScores with common methods for printing and getting
the total score. KuehlmakScores implements EvalScores.

Added reference to the layout and model to KuehlmakScores in order to
access the layout and model parameters during printing. Also added
board_type to struct KuehlmakModel.

Implemented the write method for KuehlmakModel to print a summary of
most scores, the heatmap and per-finger effort and travel scores.

In main.rs, use bench_text.json to generate more realistic scores.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>