Initial port of Simulated Annealing algorithm
Uses an iterator interface to get the next best-scoring layout. Each
iteration returns an EvalScore item.

Each iteration has a maximum number of steps. If no better solution is
found, the previous best solution is returned again.

The noise/temperature profile is adaptive to the annealing progress in
two ways:

1. Finding a new best solution is exothermic (increases the noise)
2. The rate of noise adjustment is increased when no better solution is
   found in an iteration, and decreased when a new solution is found

The maximum number of steps is the only parameter provided by the user.

TODO:
* Adjust precision
* Implement missing mutation functions
* Multi-threading

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>