eval: Add reference layout constraint
Add a constraint based on the distance of the layout to a reference
layout. Distance is measured by the number of symbols on the same key,
finger or hand. In addition to the reference layout, the constraint has
two parameters; a threshold and a weight. The threshold is the minimum
distance that causes the constraint to affect the result. If the weight
is high enough, this causes the annealing process to settle into a
layout at that defined minimum distance from the reference layout.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>