Enable ref-layout parsing in config
Enable serdes of ref-layouts in the constraint params. It can be
specified inline or as a filename of a separate layout file.

Remove setting of the ref-layout from the command line. The anneal
subcommand keeps a layout parameter for the intial layout, which may now
be different from the ref-layout used in that constraint.

Remove alphabet filtering from the choose subcommand because it no
longer has a layout parameter, and different layouts may have different
alphabets.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>