eval: Implement auto-mirroring
Automatically mirror layouts from symmetric models to avoid cluttering
the results with mirrored but otherwise identical layouts. A model is
symmetrical if a mirrored layout has the same scores. This is the case
for models with no constraints on Ortho keyboards.

Automirroring places the '.' on the right hand side of the keyboard,
which matches most common layouts, except Dvorak.

Mirroring happens lazily, when a layout is requested from the scores,
e.g. for printing or storing to a database.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>