New ranking subcommand
Rank layouts  by one or more component scores. This is a quick way to
choose a good layout out of a large set. By using ranks instead of
scores, it doesn't need to make any assumptions about the significance
or weight of different component scores. It tends to favour layouts that
balance multiple component scores and avoids layouts with unfavourable
trade-offs.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>