eval: Tweaks for column-staggered layouts
Add two column-staggered layouts (based on Ortho and Hex). Shift the
weights so that top and bottom row are about even. Add more fast bigrams
that work well on such layouts.

Also tweaked weights and key distances for Hex layouts to favor the
outer pinky in the top row as it's actually more comfortable to reach
with splay.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>