eval: Fix and improve n-gram classification
Only count dSFBs if the middle key is on a different finger. Otherwise
it's already counted as two SFBs and shouldn't be counted again. This
makes the same-finger check for travel-score correction redundant.

Only count dScissors if the middle key is not on the same finger as the
first or last key.

Do the simple checks based on hands and fingers first before the more
expensive binary searches.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>