eval: Rework roll n-grams
Rename rolls and strains to down-rolls and up-rolls. The terms "up" and
"down" contrast with "in" and "out". They are not strictly about the
direction but about comfort and strain. Down-rolls are more comfortable
than up-rolls, but up-rolls are only slightly worse.

Rename samehands to 3-rolls and change their definition. They can be
composed of both down and up-rolls. There is no need to distinguish those
3-grams, as the different value of down and up-rolls is handled by the
bigram scores.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>