eval: Count left/right-hand bigrams separately
Count left and right-hand bigrams separaetly. Calculate the score as the
sqrt of the sum of squares to promote balanced layouts. Print the sum of
left and right-hand bigrams with an indicator of the imbalance using
five buckets: worse than 75:25 or worse than 60:40 on either side, or
better than 60:40 in the middle.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>