eval: Rescale effort, travel and total scores
Make everything per 1000 key strokes. Print raw effort in addition to the
real score that uses the root of sum of squares. Normalize effort so that
a balanced score is equal ot the raw some of key costs.

Print imbalance in a more compact way with an indicator of which side
does more work.

Rebalance weights: lower weight for effort to compensate for different
effort normalization, adjust n-gram weights.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>