eval: Set affinity of Space to one or Any thumb
If Space is assigned to one thumb, e.g. on a split keyboard with Space
only on one side, that changes the load on that hand significantly. That
hand will have more key strokes, and the thumb will get involved in
rolls, redirects and contortions on that hand.

On a keyboard with a shared space-bar or two space keys, space can be hit
by Any thumb. This can avoid the unbalanced load and most of the same
hand n-grams and skipgrams. Space can always break up same-hand
sequences. It is of course up to the typist to alternate thumbs on the
shared space bar appropriately, but I choose to make some simplifying
assumptions.

Added a space_thumb parameter to KuehlmakParams to control this
behaviour.

The layout print shows the affinity of Space by shifting the key left or
right or leaving it in the middle, and adding a +-sign on one side of
the count/cost.

Changed Hand and Finger into enums and removed the corresponding consts.
Added a single "Thumb" finger to the enum. This changed array sizes for
fingers.

Other small fixes:
- use count and not cost to determine the direction of imbalance
- never treat bigrams with thumbs as LSBs
- fixed thumb travel calculations, the thumb should never travel
- use KeyProps::hand instead of key % 10 / 5 to figure out the hand

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>