Improve movement through search space
Move mutate implementation into the keyboard model, so it can account
for model-specific finger assignment. Removed operations that are not
beneficial. The best recipe seems to come down to a mix of swapping
individual keys and swapping fingers. It does slightly better on Median
and IQR than the baseline of using only key swaps.

Reimplemented finger swapping to use the actual finger assignment.
Support swapping fingers with different numbers of keys by swapping a
smaller number of keys with a random offset.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>