main: Improve estimation of expected population
A ramdom draw from a finite population is not a good model for how the
simulated annealing algorithm finds its solutions. A better estimate is
achieved by splitting the set of unique solutions found by popularity
and estimating the popularity for each fraction separately. The lowest
quartile can be mostly ignored. It can grow practically indefinitely,
but it's not worth waiting for.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>