Only store extra layout stats when needed
Extra layout stats are only needed when writing results to the database
or by kuehlmak eval -v. In all other cases that evaluate the layout with
full precision, the extra information is unnecessary.

This doesn't make a significant difference for annealing, which was
mostly using reduced precision and didn't store the extra information
anyways. But it is nearly a 2x speedup for the rank and stats
subcommands when evaluating large numbers of layouts.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>